feat: bump babel eslint deps & update settings
BREAKING CHANGE: Migrated from deprecated babel plugins to
the new @babel monorepo versions. Because of this, all of
the babel rules are now prexifixed with '@babel/<rule-name>'
instead of 'babel/<rule-name>'.

Also, a number of babel rules were deprecated so they may give
errors or warnings when running eslint. The only rules being
re-set are:

- `@babel/new-cap`
- `@babel/no-invalid-this`
- `@babel/no-unused-expressions`
- `@babel/object-curly-spacing`
- `@babel/semi`